Use grpc interceptor to gather metrics instead of explict
