+++ backend/src/test/java/ai/verta/modeldb/CollaboratorTest.java
@@ -6,12 +6,7 @@ import ai.verta.common.CollaboratorTypeEnum;
 import ai.verta.common.CollaboratorTypeEnum.CollaboratorType;
 import ai.verta.common.EntitiesEnum.EntitiesTypes;
 import ai.verta.modeldb.ProjectServiceGrpc.ProjectServiceBlockingStub;
-import ai.verta.modeldb.authservice.AuthService;
-import ai.verta.modeldb.authservice.AuthServiceUtils;
-import ai.verta.modeldb.authservice.PublicAuthServiceUtils;
-import ai.verta.modeldb.authservice.PublicRoleServiceUtils;
-import ai.verta.modeldb.authservice.RoleService;
-import ai.verta.modeldb.authservice.RoleServiceUtils;
+import ai.verta.modeldb.authservice.*;
 import ai.verta.modeldb.cron_jobs.CronJobUtils;
 import ai.verta.modeldb.cron_jobs.DeleteEntitiesCron;
 import ai.verta.modeldb.utils.ModelDBHibernateUtil;
@@ -106,7 +101,7 @@ public class CollaboratorTest {
     ModelDBHibernateUtil.runLiquibaseMigration(databasePropMap);
     App.initializeServicesBaseOnDataBase(
         serverBuilder, databasePropMap, propertiesMap, authService, roleService);
-    serverBuilder.intercept(new ModelDBAuthInterceptor());
+    serverBuilder.intercept(new AuthInterceptor());
 
     Map<String, Object> testUerPropMap = (Map<String, Object>) testPropMap.get("testUsers");
     if (testUerPropMap != null && testUerPropMap.size() > 0) {
@@ -544,7 +539,8 @@ public class CollaboratorTest {
       GetCollaborator.Response getCollaboratorResponse =
           collaboratorServiceStub.getProjectCollaborators(getCollaboratorRequest);
 
-      List<GetCollaboratorResponseItem> sharedUserList = getCollaboratorResponse.getSharedUsersList();
+      List<GetCollaboratorResponseItem> sharedUserList =
+          getCollaboratorResponse.getSharedUsersList();
       LOGGER.info(
           "Founded collaborator users count : " + getCollaboratorResponse.getSharedUsersCount());
       assertEquals(
@@ -829,7 +825,8 @@ public class CollaboratorTest {
       GetCollaborator.Response getCollaboratorResponse =
           collaboratorServiceStub.getRepositoryCollaborators(getCollaboratorRequest);
 
-      List<GetCollaboratorResponseItem> sharedUserList = getCollaboratorResponse.getSharedUsersList();
+      List<GetCollaboratorResponseItem> sharedUserList =
+          getCollaboratorResponse.getSharedUsersList();
       LOGGER.info(
           "Found collaborator users count : " + getCollaboratorResponse.getSharedUsersCount());
       assertEquals(
